Value::from_heapcell()
